+++ core/src/routes/(editor)/+layout.svelte
@@ -1,6 +1,12 @@
 <script>
 export const prerender = true;
 </script>
+
+<svelte:head>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.3/font/bootstrap-icons.css">
+    <title>Atom Page Editor</title>
+</svelte:head>
+
 <slot>
     
 </slot>