+++ core/src/routes/globals/globalstores.js
@@ -103,4 +103,36 @@ export let globalEditorViewStore = writable(EditorViews.PAGE);
 *
 * @type {Writable<String>}
 */
-export let globalSelectedElementStore = writable("");
+export let globalSelectedElementUuidStore = writable("");
+
+/**
+* Global variable which holds selected element itself in Editor panel.
+* @global
+*
+* @type {Writable<HTMLElement>}
+*/
+export let globalSelectedElementStore = writable(null);
+
+
+/**
+* Global variable which holds latter selected element itself in Editor panel.
+* @global
+*
+* @type {Writable<HTMLElement>}
+*/
+export let globalSelectedLatterElementStore = writable(null);
+
+
+/**
+* Global variable which holds latter selected element itself in Editor panel.
+* @global
+*
+* @property maxZIndex,
+* @property minZIndex,
+*
+* @type {Writable<JSON>}
+*/
+export let globalSelectedLatterElementDataStore = writable({
+    "maxZIndex": 0,
+    "minZIndex": 0
+});