+++ core/src/routes/menu/layers/layerstree.svelte
@@ -2,7 +2,7 @@
     export let component;
     export let level = 0;
 
-    import { globalSelectedElementStore } from "../../globals/globalstores.js";
+    import { globalSelectedElementUuidStore } from "../../globals/globalstores.js";
 
 </script>
 
@@ -14,7 +14,7 @@
     {#each component?.children ?? [] as child}
     
             <svelte:self component={child} level={level + 1}>
-                <span class="flex flex-row gap-2 align-middle my-2" class:text-teal-600 = {$globalSelectedElementStore === child.uuid} >
+                <span class="flex flex-row gap-2 align-middle my-2" class:text-teal-600 = {$globalSelectedElementUuidStore === child.uuid} >
                     <span class="">{child.type}</span>
                     <span class="flex flex-col">
                         <span class="text-[8px] align-middle">{child.uuid}</span><br/>