+++ public/atompagebuilder/app/_app/immutable/chunks/2-09155f09.js
@@ -1,0 +1,1 @@
+import{default as t}from"../components/pages/(app)/_layout.svelte-c137471b.js";export{t as component};