+++ public/atompagebuilder/app/_app/immutable/chunks/4-4f58aa84.js
@@ -1,0 +1,1 @@
+import{default as t}from"../components/pages/(app)/_page.svelte-a8f9248b.js";export{t as component};